requirements.txt
# Core dependencies
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.0
numpy>=1.24.0
pandas>=2.0.0

# Text processing
tiktoken>=0.5.0
nltk>=3.8.0

# Vector stores
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Clustering
scikit-learn>=1.3.0
hdbscan>=0.8.33

# Embeddings
sentence-transformers>=2.2.0

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0

# Frontend (if using Next.js)
# next>=14.0.0
# react>=18.0.0
# react-dom>=18.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.66.0

# Database
supabase>=2.0.0

# LangChain for orchestration
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.6
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0

# OpenTelemetry for observability
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-langchain>=0.1.0

